Convert solidity FixedPart.channelNonce to uint64
uint64 is a type supported by golang (as well as many other languages).
uint48 is not a golang type. Using a solidity type that is not supported
in golang can result in mismatch between data provided to an ethereum function
call parameter and the expectation for that data.

Specifically for channelNonce, with uint48, some of the calls to solidity
functions supplied integers that overflowed the type. The error message
is a cryptic "transaction reverted without a reason string". One way to
avoid the error is to match golang types for variables with solidity types.

Increasing uint48 to uint64 has minimal gas implications (and in fact results
in certain gas improvements) since channelNonce is not persisted to storage.

+++ nitro-protocol/contracts/interfaces/INitroTypes.sol
@@ -31,7 +31,7 @@ interface INitroTypes {
     struct FixedPart {
         uint256 chainId;
         address[] participants;
-        uint48 channelNonce;
+        uint64 channelNonce;
         address appDefinition;
         uint48 challengeDuration;
     }
